# backend/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Zdravdom API
  description: Home healthcare platform API
  version: 1.0.0
  contact:
    name: Zdravdom Development Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaymentStatus:
      type: string
      enum: [PENDING, PAID, REFUNDED, FAILED]
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid request data
        field:
          type: string
          example: email
        timestamp:
          type: string
          format: date-time
...
    RegisterRequest:
      type: object
      required:
        - email
        - phone
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token (15 min expiry)
        refreshToken:
          type: string
          description: JWT refresh token (7 days expiry)
        expiresIn:
          type: integer
          example: 900
...
    CreatePaymentIntentRequest:
      type: object
      required:
        - bookingId
      properties:
        bookingId:
          type: string
          format: uuid
    PaymentIntentResponse:
      type: object
      properties:
        clientSecret:
          type: string
        paymentIntentId:
          type: string
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingId:
          type: string
          format: uuid
        amount:
          type: number
          format: float
        currency:
          type: string
          default: EUR
        status:
          $ref: '#/components/schemas/PaymentStatus'
        stripePaymentIntentId:
          type: string
        createdAt:
          type: string
          format: date-time
    RefundRequest:
      type: object
      required:
        - bookingId
        - amount
      properties:
        bookingId:
          type: string
          format: uuid
        amount:
          type: number
          format: float
        reason:
          type: string

    # ─── Notifications ──────────────────────────────────────────────────────────
...
security:
  - BearerAuth: []

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /auth/register:
    post:
      summary: Register a new patient account
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Patient registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: Login with email and password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Account suspended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      summary: Refresh access token
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Patients ────────────────────────────────────────────────────────────────
...
  /payments/intent:
    post:
      summary: Create Stripe payment intent
      tags: [Payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentIntentRequest'
      responses:
        '201':
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /payments/confirm:
    post:
      summary: Confirm payment (Stripe webhook handler)
      tags: [Payments]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                paymentIntentId:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          description: Payment failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /payments/history:
    get:
      summary: Payment history
      tags: [Payments]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated payment history
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payment'
                  page:
                    type: integer
                  size:
                    type: integer
                  total:
                    type: integer
                  totalPages:
                    type: integer
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /payments/refund:
    post:
      summary: Request a refund
      tags: [Payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequest'
      responses:
        '200':
          description: Refund processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          description: Refund not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Notifications ─────────────────────────────────────────────────────────────
...